wines - sdk/requirements.txt
pyrebase4
flask
time
pandas
matplotlib